+++ controlplane/telemetry/cmd/data-api/main.go
@@ -17,9 +17,9 @@ import (
 	"github.com/malbeclabs/doublezero/controlplane/telemetry/internal/data"
 	devicedata "github.com/malbeclabs/doublezero/controlplane/telemetry/internal/data/device"
 	inetdata "github.com/malbeclabs/doublezero/controlplane/telemetry/internal/data/internet"
-	"github.com/malbeclabs/doublezero/controlplane/telemetry/pkg/epoch"
 	"github.com/malbeclabs/doublezero/smartcontract/sdk/go/serviceability"
 	"github.com/malbeclabs/doublezero/smartcontract/sdk/go/telemetry"
+	"github.com/malbeclabs/doublezero/tools/solana/pkg/epoch"
 )
 
 func main() {
