+++ controlplane/telemetry/internal/data/cli/device.go
@@ -16,9 +16,9 @@ import (
 	solanarpc "github.com/gagliardetto/solana-go/rpc"
 	"github.com/malbeclabs/doublezero/config"
 	devicedata "github.com/malbeclabs/doublezero/controlplane/telemetry/internal/data/device"
-	"github.com/malbeclabs/doublezero/controlplane/telemetry/pkg/epoch"
 	"github.com/malbeclabs/doublezero/smartcontract/sdk/go/serviceability"
 	"github.com/malbeclabs/doublezero/smartcontract/sdk/go/telemetry"
+	"github.com/malbeclabs/doublezero/tools/solana/pkg/epoch"
 	"github.com/olekukonko/tablewriter"
 	"github.com/spf13/cobra"
 )
