+++ controlplane/telemetry/internal/data/cli/internet.go
@@ -15,9 +15,9 @@ import (
 	"github.com/malbeclabs/doublezero/config"
 	internetdata "github.com/malbeclabs/doublezero/controlplane/telemetry/internal/data/internet"
 	"github.com/malbeclabs/doublezero/controlplane/telemetry/internal/data/stats"
-	"github.com/malbeclabs/doublezero/controlplane/telemetry/pkg/epoch"
 	"github.com/malbeclabs/doublezero/smartcontract/sdk/go/serviceability"
 	"github.com/malbeclabs/doublezero/smartcontract/sdk/go/telemetry"
+	"github.com/malbeclabs/doublezero/tools/solana/pkg/epoch"
 	"github.com/olekukonko/tablewriter"
 	"github.com/spf13/cobra"
 )
